demo/extract_garment/requirements.txt
gradio==4.44.1
pillow
tryondiffusion